+++ plugin-build/src/main/kotlin/io/sentry/android/gradle/util/Versions.kt
@@ -10,6 +10,18 @@ internal object AgpVersions {
     val VERSION_7_0_0: SemVer = SemVer.parse("7.0.0")
 }
 
+internal object SentryVersions {
+    internal val VERSION_DEFAULT = SemVer()
+    internal val VERSION_PERFORMANCE = SemVer(4, 0, 0)
+    internal val VERSION_OKHTTP = SemVer(5, 0, 0)
+    internal val VERSION_FILE_IO = SemVer(5, 5, 0)
+}
+
+internal object SentryModules {
+    internal const val SENTRY_ANDROID_CORE = "sentry-android-core"
+    internal const val SENTRY_ANDROID_OKHTTP = "sentry-android-okhttp"
+}
+
 /**
  * Adapted from https://github.com/swiftzer/semver/blob/master/src/main/java/net/swiftzer/semver/SemVer.kt
  */